use comprator, MeasureMetaDataRow and add back MeasureMetaDataTable.scss
