+++ src/components/editMeasure/details/MeasureMetaDataTable.scss
@@ -1,0 +1,65 @@
+$border-color: #8c8c8c;
+
+#measure-meta-data-table {
+  > .top-row {
+    display: flex;
+    flex-direction: row-reverse;
+    margin-bottom: 32px;
+  }
+
+  .meta-data-table {
+    min-width: 100%;
+
+    > thead {
+      font-weight: 500;
+      font-size: 14px;
+      background-color: #ededed;
+      border-bottom: solid 1px $border-color;
+      text-align: left;
+    }
+
+    tbody > tr {
+      border-bottom: solid 1px $border-color;
+    }
+
+    tbody > tr:last-of-type {
+      border-bottom: solid 1px transparent;
+    }
+
+    th,
+    td {
+      padding: 1.2rem 1.4rem;
+    }
+
+    tr > .col-header {
+      color: #2d3540;
+      font-family: Rubik, sans-serif;
+      font-size: 14.4px;
+      font-weight: 500;
+      text-align: left;
+    }
+  }
+
+  tbody > p {
+    text-align: center;
+    padding: 1.2rem;
+  }
+}
+
+.pagination-container {
+  display: flex;
+  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
+  border: 1px solid rgba(151, 151, 151, 0.17);
+  border-width: 0 1px 1px 1px;
+  border-top: solid 1px #8c8c8c;
+}
+
+#add-measure-meta-data-dialog {
+  .MuiButtonBase-root > svg {
+    color: #d92f2f;
+  }
+
+  .MuiFormControl-root:first-of-type {
+    margin-bottom: 24px;
+  }
+}
